need to create ne name for UniversalObject so it can fit it purpose other Objects are ok {changed the heritage of Object and created RGBContainer}

+++ Object.cc
@@ -3,13 +3,13 @@
 Object::Object (void)
 : id_() {}
 
-Object::Object (const long long unsigned& id)
+Object::Object (const uint64_t& id)
 : id_(id) {}
 
-void Object::setId (const long long unsigned& id) {
+void Object::setId (const uint64_t& id) {
 	id_ = id;
 }
 
-long long unsigned Object::getId (void) const {
+uint64_t Object::getId (void) const {
 	return id_;
 }
